Major rewrite: AQI + NowCast
I now have a better understanding if AQI, and see why it's something
that needs to be a "once/day" calculation. To that end, I've added
NowCast calculations, which require keeping track of the hourly
averages. This actually paves the way for storing the values in
flash so they can be restored across reboots and speed up
re-calculation (though we'll need to track the hours that each avg
comes from so we can age out too-old values).

- Rewrite AQI calculation based on official standards.
  - No more separate 2.5 and 10 variants. AQI is AQI, and there is no
    reason to provide the separate pseudo-AQI calculations. Plus, it
    was a lot of duplcated code for little benefit.
  - Document why we use 18 hours not 24: the official spec allows for
    calculations once we hit 75% data availability.
  - Note: This calculation still differs from the official AQI , which
    is apparently calculated midnight-to-midnight rather than a 24h
    rolling average.
- Add NowCast. This is a 12h version of AQI weighted to the most recent
  hour, and is a better fit for what a lot of people want from AQI: a
  measurement of air quality during short-term quality issues like
  wildfire smoke.
- Add sensors to report on the number of minutes remaining before AQI
  and NowCast are available.
- Update some sensors to report 1 accuracy decimal instead of 0.
- Changes around a few icons.
- Fixes some math and timing values.
- More inline comments/documentation.