Rethink AQI/NowCast calculations again
Tests identified that things just weren't firing at the right time and
AQI calculations failed to manifest when they were expected.

It turns out the interval timers and sensor updates don't correspond
consistently with each other, so this has been rewritten to account for
that by:

- Attempting calculations after every new value, rather than relying on
  an interval that may (or more often, may not) line up with the
  availability of sensor data.
- Updating scripts to be able to handle missing pm2.5 or pm10 data,
  and using `mode: restart` to allow us to simply trigger the
  calculations after each sensor value without worrying about needing
  one or both values before attempting the calculation. In this case,
  "the last write wins," and the last write (script execution) will be
  the most likely to have both values available.
- Updating "minutes remaining" when we fail to be able to calculate a
  value. This is mostly just a quality of life improvement to avoid
  situations where we've counted down to zero but still don't have
  sensor data. This could still happen, but that window is smaller.

I've also added some log messages, which may or not be helpful (but they
are for debugging).